orders_service/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.4
python-dotenv==1.0.1
django-filter==24.1
drf-yasg==1.21.7
gunicorn==21.2.0
whitenoise==6.6.0
drf-nested-routers==0.93.4